+++ src/main/kotlin/app/revanced/patches/youtube/ad/general/bytecode/patch/HideAdsPatch.kt
@@ -1,47 +1,32 @@
 package app.revanced.patches.youtube.ad.general.bytecode.patch
 
 import app.revanced.extensions.findMutableMethodOf
-import app.revanced.extensions.toErrorResult
+import app.revanced.extensions.injectHideViewCall
 import app.revanced.patcher.annotation.Description
 import app.revanced.patcher.annotation.Name
 import app.revanced.patcher.annotation.Version
 import app.revanced.patcher.data.BytecodeContext
-import app.revanced.patcher.extensions.addInstruction
-import app.revanced.patcher.extensions.instruction
 import app.revanced.patcher.patch.BytecodePatch
 import app.revanced.patcher.patch.PatchResult
 import app.revanced.patcher.patch.PatchResultSuccess
 import app.revanced.patcher.patch.annotations.DependsOn
 import app.revanced.patcher.patch.annotations.Patch
-import app.revanced.patcher.util.proxy.mutableTypes.MutableMethod
 import app.revanced.patches.shared.misc.fix.verticalscroll.patch.VerticalScrollPatch
-import app.revanced.patches.youtube.ad.general.annotation.GeneralAdsCompatibility
-import app.revanced.patches.youtube.ad.general.bytecode.fingerprints.ReelConstructorFingerprint
-import app.revanced.patches.youtube.ad.general.resource.patch.GeneralAdsResourcePatch
+import app.revanced.patches.youtube.ad.general.annotation.HideAdsCompatibility
+import app.revanced.patches.youtube.ad.general.resource.patch.HideAdsResourcePatch
 import app.revanced.patches.youtube.misc.fix.backtoexitgesture.patch.FixBackToExitGesturePatch
-import org.jf.dexlib2.iface.instruction.TwoRegisterInstruction
 import org.jf.dexlib2.iface.instruction.formats.Instruction31i
 import org.jf.dexlib2.iface.instruction.formats.Instruction35c
 
 
 @Patch
-@DependsOn([GeneralAdsResourcePatch::class, VerticalScrollPatch::class, FixBackToExitGesturePatch::class])
-@Name("general-ads")
+@DependsOn([HideAdsResourcePatch::class, VerticalScrollPatch::class, FixBackToExitGesturePatch::class])
+@Name("hide-ads")
 @Description("Removes general ads.")
-@GeneralAdsCompatibility
+@HideAdsCompatibility
 @Version("0.0.1")
-class GeneralAdsPatch : BytecodePatch(
-    listOf(ReelConstructorFingerprint)
-) {
+class HideAdsPatch : BytecodePatch() {
     override fun execute(context: BytecodeContext): PatchResult {
-        fun String.buildHideCall(viewRegister: Int) = "invoke-static { v$viewRegister }, " +
-                "Lapp/revanced/integrations/patches/GeneralAdsPatch;" +
-                "->" +
-                "$this(Landroid/view/View;)V"
-
-        fun MutableMethod.injectHideCall(insertIndex: Int, viewRegister: Int, method: String) =
-            this.addInstruction(insertIndex, method.buildHideCall(viewRegister))
-
         context.classes.forEach { classDef ->
             classDef.methods.forEach { method ->
                 with(method.implementation) {
@@ -49,7 +34,7 @@ class GeneralAdsPatch : BytecodePatch(
                         if (instruction.opcode != org.jf.dexlib2.Opcode.CONST)
                             return@forEachIndexed
                         // Instruction to store the id adAttribution into a register
-                        if ((instruction as Instruction31i).wideLiteral != GeneralAdsResourcePatch.adAttributionId)
+                        if ((instruction as Instruction31i).wideLiteral != HideAdsResourcePatch.adAttributionId)
                             return@forEachIndexed
 
                         val insertIndex = index + 1
@@ -64,25 +49,18 @@ class GeneralAdsPatch : BytecodePatch(
                             context.proxy(classDef)
                                 .mutableClass
                                 .findMutableMethodOf(method)
-                                .injectHideCall(insertIndex, viewRegister, "hideAdAttributionView")
+                                .injectHideViewCall(
+                                    insertIndex,
+                                    viewRegister,
+                                    "Lapp/revanced/integrations/patches/components/AdsFilter;",
+                                    "hideAdAttributionView"
+                                )
                         }
                     }
                 }
             }
         }
 
-        ReelConstructorFingerprint.result?.let {
-            // iput-object v$viewRegister, ...
-            val insertIndex = it.scanResult.patternScanResult!!.startIndex + 2
-
-            it.mutableMethod.apply {
-                val viewRegister = instruction<TwoRegisterInstruction>(insertIndex).registerA
-
-                injectHideCall(insertIndex, viewRegister, "hideReelView")
-            }
-        } ?: return ReelConstructorFingerprint.toErrorResult()
-
         return PatchResultSuccess()
     }
-
 }